+++ config/babel.js
@@ -8,4 +8,10 @@ module.exports = {
 	extends: [
 		'../rules/plugins/babel',
 	].map(require.resolve),
+	rules: {
+		'new-cap': 'off',
+		'object-curly-spacing': 'off',
+		'no-invalid-this': 'off',
+		semi: 'off',
+	},
 };