+++ config/typescript-mocha.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+	extends: [
+		'./typescript',
+	].map(require.resolve),
+	rules: {
+		'typescript/explicit-function-return-type': 'off',
+	},
+};