+++ rules/plugins/promise.js
@@ -3,14 +3,16 @@
 module.exports = {
 	rules: {
 		'promise/always-return': 'error',
+		'promise/avoid-new': 'off',
+		'promise/catch-or-return': 'error',
+		'promise/no-callback-in-promise': 'off',
+		'promise/no-native': 'off',
+		'promise/no-nesting': 'off',
+		'promise/no-new-statics': 'error',
+		'promise/no-promise-in-callback': 'off',
+		'promise/no-return-in-finally': 'error',
 		'promise/no-return-wrap': 'error',
 		'promise/param-names': 'error',
-		'promise/catch-or-return': 'error',
-		'promise/no-native': 'error',
-		'promise/no-nesting': 'off',
-		'promise/no-promise-in-callback': 'off',
-		'promise/no-callback-in-promise': 'off',
-		'promise/avoid-new': 'off',
-		'promise/no-return-in-finally': 'error',
+		'promise/valid-params': 'error',
 	},
 };