+++ rules/variables.js
@@ -3,7 +3,6 @@
 module.exports = {
 	rules: {
 		'init-declarations': 'off',
-		'no-catch-shadow': 'off',
 		'no-delete-var': 'off',
 		'no-label-var': 'error',
 		'no-restricted-globals': 'off',